Adds menu inside Courses and includes the back arrow menu navigation to Course, Dashboard, and User
